feat: Add distributed tracing with Zipkin
This commit introduces distributed tracing to the project using Zipkin.

- Adds a `docker-compose.yml` file to easily start a Zipkin instance.
- Includes the `micrometer-tracing-bridge-brave` and `zipkin-reporter-brave` dependencies in all microservices (`api-gateway`, `config-server`, `eureka-server`, `security-module`).
- Configures each microservice's `application.yml` to send traces to the Zipkin endpoint `http://localhost:9411/api/v2/spans` and sets the sampling probability to 1.0.
- Updates the `README.md` to include instructions on how to start the Zipkin server using Docker Compose.

This will allow developers to visualize the flow of requests between services, making it easier to debug and understand the system's behavior.
