--- api-gateway/src/main/java/com/example/apigateway/ApiGatewayApplication.java
@@ -1,15 +1,0 @@
-package com.example.apigateway;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class ApiGatewayApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ApiGatewayApplication.class, args);
-    }
-
-}
