--- api-gateway/src/main/resources/application.properties
@@ -1,6 +1,0 @@
-spring.application.name=api-gateway
-server.port=8080
-
-eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
-
-spring.cloud.gateway.discovery.locator.enabled=true
