+++ config-server/src/main/resources/application.yml
@@ -1,0 +1,20 @@
+server:
+  port: '8888'
+spring:
+  application:
+    name: config-server
+  profiles:
+    active: git
+  cloud:
+    config:
+      server:
+        git:
+          uri: https://github.com/david-eve-za/config-repository
+
+management:
+  tracing:
+    sampling:
+      probability: 1.0
+  zipkin:
+    tracing:
+      endpoint: http://localhost:9411/api/v2/spans
