--- eureka-server/src/main/resources/application.properties
@@ -1,4 +1,0 @@
-server.port=8761
-
-eureka.client.register-with-eureka=false
-eureka.client.fetch-registry=false
