+++ eureka-server/src/main/resources/application.yml
@@ -1,0 +1,19 @@
+server:
+  port: '8761'
+
+spring:
+  application:
+    name: eureka-server
+
+eureka:
+  client:
+    register-with-eureka: 'false'
+    fetch-registry: 'false'
+
+management:
+  tracing:
+    sampling:
+      probability: 1.0
+  zipkin:
+    tracing:
+      endpoint: http://localhost:9411/api/v2/spans
