+++ eureka-server/src/main/resources/bootstrap.yml
@@ -1,0 +1,6 @@
+spring:
+  application:
+    name: eureka-server
+  cloud:
+    config:
+      uri: http://localhost:8888
