+++ pom.xml
@@ -45,6 +45,7 @@
     <module>eureka-server</module>
     <module>api-gateway</module>
     <module>security-module</module>
+    <module>config-server</module>
   </modules>
 
   <build>
