--- security-module/src/main/java/com/example/security/SecurityModuleApplication.java
@@ -1,13 +1,0 @@
-package com.example.security;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SecurityModuleApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SecurityModuleApplication.class, args);
-    }
-
-}
