--- security-module/src/main/java/com/example/security/config/DataInitializer.java
@@ -1,23 +1,0 @@
-package com.example.security.config;
-
-import com.example.security.model.User;
-import com.example.security.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class DataInitializer implements CommandLineRunner {
-
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    @Override
-    public void run(String... args) {
-        if (userRepository.findByUsername("user").isEmpty()) {
-            userRepository.save(new User(null, "user", passwordEncoder.encode("password")));
-        }
-    }
-}
