--- security-module/src/main/java/com/example/security/model/AuthRequest.java
@@ -1,18 +1,0 @@
-package com.example.security.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * Represents an authentication request containing user credentials.
- */
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthRequest {
-
-    private String username;
-    private String password;
-
-}
