--- security-module/src/main/java/com/example/security/model/AuthResponse.java
@@ -1,12 +1,0 @@
-package com.example.security.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthResponse {
-    private String token;
-}
