--- security-module/src/main/java/com/example/security/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.example.security.repository;
-
-import com.example.security.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-}
