--- security-module/src/main/java/com/example/security/util/JwtUtil.java
@@ -1,129 +1,0 @@
-package com.example.security.util;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-
-import java.security.Key;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.function.Function;
-
-/**
- * Utility class for handling JSON Web Tokens (JWT).
- * Provides methods for generating, validating, and extracting information from JWTs.
- */
-@Component
-public class JwtUtil {
-
-    @Value("${jwt.secret}")
-    private String SECRET_KEY;
-
-    @Value("${jwt.expiration}")
-    private long expiration;
-
-
-    /**
-     * Extracts the username from a JWT token.
-     * @param token The JWT token.
-     * @return The username (subject) from the token.
-     */
-    public String extractUsername(String token) {
-        return extractClaim(token, Claims::getSubject);
-    }
-
-    /**
-     * Extracts the expiration date from a JWT token.
-     * @param token The JWT token.
-     * @return The expiration date from the token.
-     */
-    public Date extractExpiration(String token) {
-        return extractClaim(token, Claims::getExpiration);
-    }
-
-    /**
-     * Extracts a specific claim from a JWT token.
-     * @param token The JWT token.
-     * @param claimsResolver Function to resolve the desired claim.
-     * @param <T> The type of the claim.
-     * @return The extracted claim.
-     */
-    public <T> T extractClaim(String token, Function<Claims, T> claimsResolver) {
-        final Claims claims = extractAllClaims(token);
-        return claimsResolver.apply(claims);
-    }
-
-    /**
-     * Extracts all claims from a JWT token.
-     * @param token The JWT token.
-     * @return All claims from the token.
-     */
-    private Claims extractAllClaims(String token) {
-        return Jwts.parserBuilder().setSigningKey(getSignKey()).build().parseClaimsJws(token).getBody();
-    }
-
-    /**
-     * Checks if the JWT token is expired.
-     * @param token The JWT token.
-     * @return True if the token is expired, false otherwise.
-     */
-    private Boolean isTokenExpired(String token) {
-        return extractExpiration(token).before(new Date());
-    }
-
-    /**
-     * Validates a JWT token against user details.
-     * @param token The JWT token.
-     * @param userDetails The user details.
-     * @return True if the token is valid, false otherwise.
-     */
-    public Boolean validateToken(String token, UserDetails userDetails) {
-        try {
-            final String username = extractUsername(token);
-            return (username.equals(userDetails.getUsername()) && !isTokenExpired(token));
-        } catch (ExpiredJwtException e) {
-            return false;
-        }
-    }
-
-    /**
-     * Generates a JWT token for the given username.
-     * @param username The username for which to generate the token.
-     * @return The generated JWT token.
-     */
-    public String generateToken(String username) {
-        Map<String, Object> claims = new HashMap<>();
-        return createToken(claims, username);
-    }
-
-    /**
-     * Creates a JWT token with specified claims and subject.
-     * @param claims The claims to include in the token.
-     * @param username The subject of the token (username).
-     * @return The created JWT token.
-     */
-    private String createToken(Map<String, Object> claims, String username) {
-        return Jwts.builder()
-                .setClaims(claims)
-                .setSubject(username)
-                .setIssuedAt(new Date(System.currentTimeMillis()))
-                .setExpiration(new Date(System.currentTimeMillis() + expiration))
-                .signWith(getSignKey(), SignatureAlgorithm.HS256).compact();
-    }
-
-    /**
-     * Retrieves the signing key for JWT operations.
-     * @return The signing key.
-     */
-    private Key getSignKey() {
-        byte[] keyBytes = Decoders.BASE64.decode(SECRET_KEY);
-        return Keys.hmacShaKeyFor(keyBytes);
-    }
-}
