+++ security-module/src/main/java/gon/cue/security/SecurityModuleApplication.java
@@ -1,0 +1,13 @@
+package gon.cue.security;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SecurityModuleApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SecurityModuleApplication.class, args);
+    }
+
+}
