+++ security-module/src/main/java/gon/cue/security/config/DataInitializer.java
@@ -1,0 +1,23 @@
+package gon.cue.security.config;
+
+import gon.cue.security.model.User;
+import gon.cue.security.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class DataInitializer implements CommandLineRunner {
+
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    @Override
+    public void run(String... args) {
+        if (userRepository.findByUsername("user").isEmpty()) {
+            userRepository.save(new User(null, "user", passwordEncoder.encode("password")));
+        }
+    }
+}
