+++ security-module/src/main/java/gon/cue/security/model/AuthRequest.java
@@ -1,0 +1,18 @@
+package gon.cue.security.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * Represents an authentication request containing user credentials.
+ */
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthRequest {
+
+    private String username;
+    private String password;
+
+}
