+++ security-module/src/main/java/gon/cue/security/model/AuthResponse.java
@@ -1,0 +1,12 @@
+package gon.cue.security.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthResponse {
+    private String token;
+}
