+++ security-module/src/main/java/gon/cue/security/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package gon.cue.security.repository;
+
+import gon.cue.security.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+}
