--- security-module/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-jwt.secret=bllRbxCOXTiYhFGAapfUb4ob3bglSuz7QJ8xUpmTV9M=
-jwt.expiration=36000000
