+++ security-module/src/main/resources/bootstrap.yml
@@ -1,0 +1,6 @@
+spring:
+  application:
+    name: security-module
+  cloud:
+    config:
+      uri: http://localhost:8888
