--- security-module/src/test/java/com/example/security/service/UserDetailsServiceImplTest.java
@@ -1,48 +1,0 @@
-package com.example.security.service;
-
-import com.example.security.model.User;
-import com.example.security.repository.UserRepository;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.when;
-
-public class UserDetailsServiceImplTest {
-
-    @InjectMocks
-    private UserDetailsServiceImpl userDetailsService;
-
-    @Mock
-    private UserRepository userRepository;
-
-    @BeforeEach
-    public void setUp() {
-        MockitoAnnotations.openMocks(this);
-    }
-
-    @Test
-    void loadUserByUsername_userFound() {
-        User mockUser = new User(1L, "user", "encodedPassword");
-        when(userRepository.findByUsername("user")).thenReturn(Optional.of(mockUser));
-
-        UserDetails userDetails = userDetailsService.loadUserByUsername("user");
-        assertNotNull(userDetails);
-        assertEquals("user", userDetails.getUsername());
-        assertEquals("encodedPassword", userDetails.getPassword());
-        assertTrue(userDetails.getAuthorities().isEmpty());
-    }
-
-    @Test
-    void loadUserByUsername_userNotFound() {
-        when(userRepository.findByUsername("nonexistent")).thenReturn(Optional.empty());
-        assertThrows(UsernameNotFoundException.class, () -> userDetailsService.loadUserByUsername("nonexistent"));
-    }
-}
